app/backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.5
    # via
    #   -r requirements.in
    #   openai
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
    #   watchfiles
asgiref==3.7.2
    # via opentelemetry-instrumentation-asgi
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
azure-common==1.1.28
    # via azure-search-documents
azure-core==1.29.4
    # via
    #   azure-core-tracing-opentelemetry
    #   azure-identity
    #   azure-monitor-opentelemetry
    #   azure-monitor-opentelemetry-exporter
    #   azure-search-documents
    #   azure-storage-blob
    #   msrest
azure-core-tracing-opentelemetry==1.0.0b11
    # via azure-monitor-opentelemetry
azure-identity==1.14.0
    # via -r requirements.in
azure-monitor-opentelemetry==1.0.0
    # via -r requirements.in
azure-monitor-opentelemetry-exporter==1.0.0b17
    # via azure-monitor-opentelemetry
azure-search-documents==11.4.0b6
    # via -r requirements.in
azure-storage-blob==12.18.2
    # via -r requirements.in
certifi==2023.7.22
    # via
    #   msrest
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via uvicorn
cryptography==41.0.4
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
deprecated==1.2.14
    # via opentelemetry-api
dnspython==2.4.2
    # via pymongo
et-xmlfile==1.1.0
    # via openpyxl
exceptiongroup==1.1.3
    # via anyio
fastapi==0.104.1
    # via -r requirements.in
fixedint==0.1.6
    # via azure-monitor-opentelemetry-exporter
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via uvicorn
httptools==0.6.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
importlib-metadata==6.8.0
    # via opentelemetry-api
isodate==0.6.1
    # via
    #   azure-search-documents
    #   azure-storage-blob
    #   msrest
msal==1.24.1
    # via
    #   -r requirements.in
    #   azure-identity
    #   msal-extensions
msal-extensions==1.0.0
    # via
    #   -r requirements.in
    #   azure-identity
msrest==0.7.1
    # via azure-monitor-opentelemetry-exporter
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==1.26.0
    # via
    #   openai
    #   pandas
    #   pandas-stubs
oauthlib==3.2.2
    # via requests-oauthlib
openai[datalib]==0.28.1
    # via -r requirements.in
openpyxl==3.1.2
    # via openai
opentelemetry-api==1.20.0
    # via
    #   azure-core-tracing-opentelemetry
    #   azure-monitor-opentelemetry-exporter
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-psycopg2
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-urllib
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-instrumentation==0.41b0
    # via
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-psycopg2
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-urllib
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-aiohttp-client==0.41b0
    # via -r requirements.in
opentelemetry-instrumentation-asgi==0.41b0
    # via
    #   -r requirements.in
    #   opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-dbapi==0.41b0
    # via opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-django==0.41b0
    # via azure-monitor-opentelemetry
opentelemetry-instrumentation-fastapi==0.41b0
    # via azure-monitor-opentelemetry
opentelemetry-instrumentation-flask==0.41b0
    # via azure-monitor-opentelemetry
opentelemetry-instrumentation-psycopg2==0.41b0
    # via azure-monitor-opentelemetry
opentelemetry-instrumentation-requests==0.41b0
    # via
    #   -r requirements.in
    #   azure-monitor-opentelemetry
opentelemetry-instrumentation-urllib==0.41b0
    # via azure-monitor-opentelemetry
opentelemetry-instrumentation-urllib3==0.41b0
    # via azure-monitor-opentelemetry
opentelemetry-instrumentation-wsgi==0.41b0
    # via
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-flask
opentelemetry-resource-detector-azure==0.1.0
    # via azure-monitor-opentelemetry
opentelemetry-sdk==1.20.0
    # via
    #   azure-monitor-opentelemetry-exporter
    #   opentelemetry-resource-detector-azure
opentelemetry-semantic-conventions==0.41b0
    # via
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-urllib
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-util-http==0.41b0
    # via
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-urllib
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-instrumentation-wsgi
packaging==23.1
    # via opentelemetry-instrumentation-flask
pandas==2.1.1
    # via openai
pandas-stubs==2.1.1.230928
    # via openai
portalocker==2.8.2
    # via msal-extensions
pycparser==2.21
    # via cffi
pydantic==2.4.2
    # via fastapi
pydantic-core==2.10.1
    # via pydantic
pyjwt[crypto]==2.8.0
    # via
    #   msal
    #   pyjwt
pymongo==4.5.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.0
    # via uvicorn
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via uvicorn
regex==2023.8.8
    # via tiktoken
requests==2.31.0
    # via
    #   azure-core
    #   msal
    #   msrest
    #   openai
    #   requests-oauthlib
    #   tiktoken
requests-oauthlib==1.3.1
    # via msrest
six==1.16.0
    # via
    #   azure-core
    #   isodate
    #   python-dateutil
sniffio==1.3.0
    # via anyio
starlette==0.27.0
    # via fastapi
tiktoken==0.5.1
    # via -r requirements.in
tqdm==4.66.1
    # via openai
types-pytz==2023.3.1.1
    # via pandas-stubs
typing-extensions==4.8.0
    # via
    #   asgiref
    #   azure-core
    #   azure-storage-blob
    #   fastapi
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   uvicorn
tzdata==2023.3
    # via pandas
urllib3==2.0.6
    # via requests
uvicorn[standard]==0.23.2
    # via -r requirements.in
uvloop==0.17.0
    # via uvicorn
watchfiles==0.20.0
    # via uvicorn
websockets==11.0.3
    # via uvicorn
wrapt==1.15.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-urllib3
yarl==1.9.2
    # via aiohttp
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
